+++ src/pocketmine/entity/Entity.php
@@ -30,6 +30,16 @@ use pocketmine\block\Portal;
 
 use pocketmine\block\Water;
 use pocketmine\block\SlimeBlock;
+use pocketmine\entity\Hostile\Blaze;
+use pocketmine\entity\Hostile\Ghast;
+use pocketmine\entity\Hostile\Husk;
+use pocketmine\entity\Hostile\Silverfish;
+use pocketmine\entity\Hostile\Skeleton;
+use pocketmine\entity\Hostile\Slime;
+use pocketmine\entity\Hostile\Witch;
+use pocketmine\entity\Hostile\Wither;
+use pocketmine\entity\Hostile\Zombie;
+use pocketmine\entity\Hostile\ZombieVillager;
 use pocketmine\entity\Item as DroppedItem;
 use pocketmine\entity\Neutral\CaveSpider;
 use pocketmine\entity\Neutral\Enderman;
@@ -38,6 +48,7 @@ use pocketmine\entity\Neutral\Spider;
 use pocketmine\entity\Passive\Bat;
 use pocketmine\entity\Passive\Chicken;
 use pocketmine\entity\Passive\Cow;
+use pocketmine\entity\Passive\Horse;
 use pocketmine\entity\Passive\Mooshroom;
 use pocketmine\entity\Passive\Pig;
 use pocketmine\entity\Passive\Rabbit;
