+++ src/pocketmine/entity/Hostile/Husk.php
@@ -10,7 +10,7 @@
  * \_____/ |_____| |_|  \_| |_| /_____/  /_/     /_____/
  *
  * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published by
+ * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
  *
@@ -19,40 +19,34 @@
  *
  */
 
-namespace pocketmine\entity;
+namespace pocketmine\entity\Hostile;
 
+use pocketmine\entity\Entity;
 use pocketmine\Player;
 use pocketmine\network\protocol\AddEntityPacket;
 
-class ZombieVillager extends Zombie{
-	const NETWORK_ID = 44;
-
-	public $width = 1.031;
-	public $length = 0.891;
-	public $height = 2.125;
-
-	public function initEntity(){
-		$this->setMaxHealth(20);
-		parent::initEntity();
-	}
-
+class Husk extends Zombie{
+	const NETWORK_ID = 47;
+	
 	public function getName() : string{
-		return "Zombie Villager";
+		return "Husk";
 	}
-
+	
 	public function spawnTo(Player $player){
 		$pk = new AddEntityPacket();
-		$pk->type = ZombieVillager::NETWORK_ID;
 		$pk->eid = $this->getId();
+		$pk->type = Husk::NETWORK_ID;
 		$pk->x = $this->x;
 		$pk->y = $this->y;
 		$pk->z = $this->z;
 		$pk->speedX = $this->motionX;
 		$pk->speedY = $this->motionY;
 		$pk->speedZ = $this->motionZ;
+		$pk->yaw = $this->yaw;
+		$pk->pitch = $this->pitch;
 		$pk->metadata = $this->dataProperties;
 		$player->dataPacket($pk);
 
-		parent::spawnTo($player);
+		Entity::spawnTo($player);
 	}
 }
