+++ src/pocketmine/entity/Hostile/Skeleton.php
@@ -10,7 +10,7 @@
  * \_____/ |_____| |_|  \_| |_| /_____/  /_/     /_____/
  *
  * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
+ * it under the terms of the GNU Lesser General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
  *
@@ -19,22 +19,29 @@
  *
  */
 
-namespace pocketmine\entity;
 
-use pocketmine\Player;
+namespace pocketmine\entity\Hostile;
+
+use pocketmine\entity\Monster;
+use pocketmine\entity\ProjectileSource;
 use pocketmine\network\protocol\AddEntityPacket;
+use pocketmine\Player;
+use pocketmine\network\protocol\MobEquipmentPacket;
+use pocketmine\item\Item as ItemItem;
 
-class Husk extends Zombie{
-	const NETWORK_ID = 47;
+class Skeleton extends Monster implements ProjectileSource {
+	const NETWORK_ID = 34;
+
+	public $dropExp = [5, 5];
 	
 	public function getName() : string{
-		return "Husk";
+		return "Skeleton";
 	}
 	
 	public function spawnTo(Player $player){
 		$pk = new AddEntityPacket();
 		$pk->eid = $this->getId();
-		$pk->type = Husk::NETWORK_ID;
+		$pk->type = Skeleton::NETWORK_ID;
 		$pk->x = $this->x;
 		$pk->y = $this->y;
 		$pk->z = $this->z;
@@ -46,6 +53,14 @@ class Husk extends Zombie{
 		$pk->metadata = $this->dataProperties;
 		$player->dataPacket($pk);
 
-		Entity::spawnTo($player);
+		parent::spawnTo($player);
+		
+		$pk = new MobEquipmentPacket();
+		$pk->eid = $this->getId();
+		$pk->item = new ItemItem(ItemItem::BOW);
+		$pk->slot = 0;
+		$pk->selectedSlot = 0;
+
+		$player->dataPacket($pk);
 	}
 }
