+++ src/pocketmine/entity/Hostile/WitherSkeleton.php
@@ -2,11 +2,11 @@
 
 /*
  *
- *  ____            _        _   __  __ _                  __  __ ____  
- * |  _ \ ___   ___| | _____| |_|  \/  (_)_ __   ___      |  \/  |  _ \ 
+ *  ____            _        _   __  __ _                  __  __ ____
+ * |  _ \ ___   ___| | _____| |_|  \/  (_)_ __   ___      |  \/  |  _ \
  * | |_) / _ \ / __| |/ / _ \ __| |\/| | | '_ \ / _ \_____| |\/| | |_) |
- * |  __/ (_) | (__|   <  __/ |_| |  | | | | | |  __/_____| |  | |  __/ 
- * |_|   \___/ \___|_|\_\___|\__|_|  |_|_|_| |_|\___|     |_|  |_|_| 
+ * |  __/ (_) | (__|   <  __/ |_| |  | | | | | |  __/_____| |  | |  __/
+ * |_|   \___/ \___|_|\_\___|\__|_|  |_|_|_| |_|\___|     |_|  |_|_|
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as published by
@@ -15,28 +15,32 @@
  *
  * @author PocketMine Team
  * @link http://www.pocketmine.net/
- * 
+ *
  *
 */
 
-namespace pocketmine\entity;
+namespace pocketmine\entity\Hostile;
 
+use pocketmine\entity\Monster;
+use pocketmine\entity\ProjectileSource;
 use pocketmine\network\protocol\AddEntityPacket;
 use pocketmine\Player;
+use pocketmine\network\protocol\MobEquipmentPacket;
+use pocketmine\item\Item as ItemItem;
 
-class Silverfish extends Monster{
-	const NETWORK_ID = 39;
+class WitherSkeleton extends Monster implements ProjectileSource {
+	const NETWORK_ID = 48;
 
 	public $dropExp = [5, 5];
 	
 	public function getName() : string{
-		return "Silverfish";
+		return "Wither Skeleton";
 	}
 	
 	public function spawnTo(Player $player){
 		$pk = new AddEntityPacket();
 		$pk->eid = $this->getId();
-		$pk->type = Silverfish::NETWORK_ID;
+		$pk->type = WitherSkeleton::NETWORK_ID;
 		$pk->x = $this->x;
 		$pk->y = $this->y;
 		$pk->z = $this->z;
@@ -49,5 +53,13 @@ class Silverfish extends Monster{
 		$player->dataPacket($pk);
 
 		parent::spawnTo($player);
+		
+		$pk = new MobEquipmentPacket();
+		$pk->eid = $this->getId();
+		$pk->item = new ItemItem(ItemItem::STONE_SWORD);
+		$pk->slot = 0;
+		$pk->selectedSlot = 0;
+
+		$player->dataPacket($pk);
 	}
 }
