+++ src/pocketmine/entity/Hostile/Zombie.php
@@ -19,31 +19,38 @@
  *
 */
 
-namespace pocketmine\entity;
+namespace pocketmine\entity\Hostile;
 
+
+use pocketmine\event\entity\EntityDamageByEntityEvent;
+use pocketmine\item\Item as ItemItem;
+use pocketmine\item\enchantment\Enchantment;
 use pocketmine\network\protocol\AddEntityPacket;
 use pocketmine\Player;
+use pocketmine\entity\Monster;
 
-class Ghast extends FlyingAnimal{
-	const NETWORK_ID = 41;
+class Zombie extends Monster {
+	const NETWORK_ID = 32;
 
-	public $width = 6;
-	public $length = 6;
-	public $height = 6;
-	
+	public $width = 0.6;
+	public $length = 0.6;
+	public $height = 1.8;
+
+	public $dropExp = [5, 5];
+
 	public function getName() : string{
-		return "Ghast";
+		return "Zombie";
 	}
-
+	
 	public function initEntity(){
-		$this->setMaxHealth(10);
+		$this->setMaxHealth(20);
 		parent::initEntity();
 	}
-	
+
 	public function spawnTo(Player $player){
 		$pk = new AddEntityPacket();
 		$pk->eid = $this->getId();
-		$pk->type = Ghast::NETWORK_ID;
+		$pk->type = Zombie::NETWORK_ID;
 		$pk->x = $this->x;
 		$pk->y = $this->y;
 		$pk->z = $this->z;
@@ -56,5 +63,33 @@ class Ghast extends FlyingAnimal{
 		$player->dataPacket($pk);
 
 		parent::spawnTo($player);
+	}
+
+	public function getDrops(){
+		$lootingL = 0;
+		$cause = $this->lastDamageCause;
+		$drops = [];
+		if($cause instanceof EntityDamageByEntityEvent and $cause->getDamager() instanceof Player){
+			$lootingL = $cause->getDamager()->getItemInHand()->getEnchantmentLevel(Enchantment::TYPE_WEAPON_LOOTING);
+			if(mt_rand(0, 199) < (5 + 2 * $lootingL)){
+				switch(mt_rand(0, 3)){
+					case 0:
+						$drops[] = ItemItem::get(ItemItem::IRON_INGOT, 0, 1);
+						break;
+					case 1:
+						$drops[] = ItemItem::get(ItemItem::CARROT, 0, 1);
+						break;
+					case 2:
+						$drops[] = ItemItem::get(ItemItem::POTATO, 0, 1);
+						break;
+				}
+			}
+			$count = mt_rand(0, 2 + $lootingL);
+			if($count > 0){
+				$drops[] = ItemItem::get(ItemItem::ROTTEN_FLESH, 0, $count);
+			}
+		}
+
+		return $drops;
 	}
 }
